requirements.txt
google-generativeai
python-dotenv
rich
langgraph
langchain
langchain-google-genai
fastapi
uvicorn
streamlit
requests
duckduckgo-search
pillow
langchain-community
google-adk
